requirements.txt
Django==2.1.1
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-js-asset==1.1.0
mysqlclient==1.3.11
Pillow==5.2.0
pytz==2018.5
